backend/requirements.txt
# ============================
# AI Data Analyst - Backend
# LangChain 0.3.x + Groq (STABLE FOR AGENTS)
# ============================

# -------- Web Framework --------
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# -------- Database --------
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# -------- Authentication --------
python-jose[cryptography]>=3.3.0
passlib==1.7.4
bcrypt==3.2.2
python-multipart>=0.0.6

# -------- AI / LLM (CRITICAL - GROQ COMPATIBLE) --------
langchain==0.3.22
langchain-core==0.3.56
langchain-community==0.3.20
langchain-groq==0.2.1
python-dotenv>=1.0.0

# -------- Hybrid Search --------
rank-bm25>=0.2.2
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
scikit-learn>=1.3.0

# -------- Data Processing --------
numpy<2.0
pandas>=2.1.0,<2.2.0
numexpr>=2.8.0
bottleneck>=1.3.7

# -------- Configuration --------
python-dotenv>=1.0.0

# -------- Dev / Testing --------
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# -------- Pydantic --------
pydantic>=2.12.0
email-validator>=2.1.1
